Generate config base during build
Previously, the config base was compiled once by Vlad and uploaded to
Swift. Each subsequent installation used that pre-compiled config base.

Now, we compile the config-base at every build, and embed it in the
hcf.tar.gz file that is distributed to users.

If you do not want to run a full build, but want to deploy via
 Terraform, you need to grab a compiled hcf-config.tar.gz from Jenkins.