HFS-332 - staged droplet uploads need to be shared across multiple controllers
The resources, like cc-buildpacks, cc-resources, and cc-droplets are
currently being shared by nfs-mounting /var/vcap/nfs/shared to /data/cf-api,
which is correct. On Stackato we found that when you're running more than
one controller you also need to share
/var/*/data/cloud_controller_ng/tmp/staged_droplet_uploads , because
different controllers can serve subsequent requests as part of a push
command.  The first controller will upload droplets to that directory.
When the next controller is invoked, it can't find the droplets it's
expecting to see, and the push fails.