HCF-400 HCF-Infrastructure version pinning makes gato wrapper fail
Fetch the gato version, nearly identical to how we fetch the configgin
and fissile builds. This will fetch it from docker, and retrieve the
version from the call. This has one drawback, it's slightly fragile, but
in a way that does not impact the customer, since that version check is
at build time.

This will create a separate variable, gato-build, which contains the
gato build number.