HCF-381 - Explicitly yaml-encode multi-line strings for gato
The current version of gato doesn't yaml-encode values to be written
to consul. It only checks to see if they are yaml-decodable, and if not
faults. The problem is that a simple multi-line string is valid Yaml,
but decodes as a single-line string.  We need to explicitly yaml-encode
the string to preserve the newlines when assigned to cert clients.