[HCF-578] Increase mysql proxy timeout
The switchboard checks the health of mysql backends using an HTTP endpoint.
By default, the timeout for the GET for those endpoints is 5000 ms. If the timeout is reached, the switchboard considers the backend to be dead, and severs connections.
With the new timeout, there were no dropped connections across 100,000 "cf apps" requests, in ~8 hrs.