Set garden DNS client to Docker's internal DNS server
This commit makes the assumption that the internal DNS
server from docker is accessible on the x.x.x.1 IP address
from a docker private network.
This behavior is not explicitly documented, but it's the
simplest way to access the docker service discovery from
a garden container (which is inside a docker container).

Essentially, this will allow CF apps to resolve *.hcf hostnames
and to access CF services deployed with docker inside hcf network.