HCF-630 Add bootstrap node to avoid race on start
In a BOSH based release, the first etcd node is a 'canary' that blocks
the other nodes from starting. In the HCF setup, all nodes come up
simultaneously, so they can get in a state where each thinks it's the
master and they never cluster.

This patch uses 'bootstrap_node' to set the master node, and the other
nodes will wait until the master node is running before they join the
cluster.