HCF-627 mysql-proxy: properties.cf_mysql.proxy.proxy_ips contains hostnames
We don't really deal with IP addresses; instead, we deal in host names.
This makes it more possible to find the instance in the list so it can give
a correct index.

Once CAPS-317 is fixed, the role manifest will need to use the new
environment variables to build the list.