HCF-825 Remove ENV_DIR setting when generating configs
The environment directories were not as useful as we thought - we basically
only ever used the hard-coded variants.  Reflect that by removing the need
(and the option) to set ENV_DIR.  This also lets us merge the different
variants better (by letting rm-transformer take multiple --env-dir= params
and having a way to un-override previous settings).

This also removes the unused --no-env-dir option.