[HCFRO-375] Add version to windows cell installer
The version number received by package.ps1 will be set as the installer
version and will
be shown when the installer is run and also when version.ps1 is run.
The version is created from a base version (i.e.: 1.0.0.) and an minor
version (i.e.: 33-3b4d543).
The base version is hardcoded in the package.ps1 script.
The minor version is obtained from the Jenkins BUILD_NUMBER-(git
describe).