[HCF-914][HCF-915][HCF-916][HCF-917] Self-clustering the roles "etcd", "nats", "api", and "consul".
__Note__ currently, HCP will not cycle the existing instances when it
scales a component. E.g., start with 1 node, then manually set it to
min/max 3 instances (i.e. turn it into HA).

To work around this, we apply a dummy environment variable
`..._CONFIG_REVISION` (that will not actually be used anywhere)
describing the number of instances, so that HCP can use it to cycle
the correct components.  This relies on the user correctly changing
the variable to the number of instances (however, we don't actually
require the value to be correct, only that it changes at the
appropriate times).