HCF-913 Add MYSQL_CLUSTER_CONFIG_REVISION for HCP component cycling
Currently, HCP will not cycle the existing instances when it scales a
component.  E.g., start with 1 MySQL node, then manually set it to
min/max 3 instances (i.e. turn it into HA).  Also, when the size of the
MySQL cluster changes, we will need to cycle mysql-proxy as well.

To work around this, we apply a dummy environment variable (that will
not actually be used anywhere) describing the number of MySQL instances,
so that HCP can use it to cycle the correct components.  This relies on
the user correctly changing the variable to the number of instances
(however, we don't actually require the value to be correct, only that
it changes at the appropriate times).