HCF-977 Unset proxies before creating UAA users
The uaac client will look at $HTTP_PROXY to figure out what proxy it
should use, but it does _not_ obey $NO_PROXY which causes it to go
through the proxy to hit UAA.  This is totally incorrect as UAA is
within the same HCP cluster, and it causes attempts to talk to UAA to
fail.  Workaround this by unsetting all the proxy-related environment
variables at the top of the file before running uaac.