HCF-869 Rewrite sso-create-service job in Ruby to use a OAuth client
We want to switch to an oauth client rather than the default user, as
the user password could be changed post-deployment (by the cluster admin
or via other means).  However, the `cf` cli doesn't support using a
client_credentials grant, so we need to switch to curl.  In order to
parse the JSON output for the access token, it is actually easier to
just switch everything over to Ruby :(