HCF-980 Mark applications as STAGED after droplets are added
Otherwise rolling back a droplet will not work if the application
is in the FAILED state; the package_state is normally only updated
during the "staging complete" message from the stager.

There should be no harm in marking the package prematurely, as non
of the app start commands will be issued by this action.