HCF-869 Convert autoscaler-register job to not use cf users
The cluster administrator might change the password of the initial admin
after cluster deployment; if the job gets re-run (e.g. because of
cluster configuration changes), it will fail due to an incorrect
password.  Convert everything to use an OAuth client instead of an user
to avoid the issue.  Unfortunately this means we will need to parse
JSON, which is easier in Ruby than shell.