HCF-964 Docs for running acceptance/smoke tests against HCF off vagrant
We can actually use all the vagrant settings, and just override the
domain when we run them.  This makes it easier than setting up all the
various parameters required, but limits us to not running it against
HSM-deployed configurations (where a lot of the settings are generated
and we don't know what the correct values are).