HCF-583 Update compile cache management for new compilation paths
We no longer have the package name in the compilation paths (because
some packages have different names across releases for the same
fingerprint).  This is a good opportunity to also switch to XZ for
compression, which is desirable as that reduces the amount of data we
need to transfer over to/from the host.