HCF-728 Disable HA Proxy
HA Proxy isn't needed as HCP will manage the load balancing for us.

Note that this means diego-access will have a public port for ssh, and
require a DNS entry for its new ELB.

The router now has the public ports, so the ELB that previously pointed
to haproxy now points to it.