HCF-1059 cc: don't run migrations in pre-start
Having things fail in prestart is a bit more expensive for us than
having the same fail under monit (because any pre-start failures cause
the container to be completely destroyed and restarted).  Adjust cloud
controller configuration to not run migrations in pre-start, and instead
do so on normal startup of (the first instance of) the cloud controller.