HCF-1089 - Speed up the time to bosh-create a release by stubbing out git.
create-release only calls `git status` (twice), to determine if the
source is dirty.  The first time we don't care, because we're running
with --force, but the code is backwards:

raise_dirty_state_error if dirty_state? && !force

The checks are wrong because it's much cheaper to check a
boolean than run `git status`.

The result of the second call to `dirty_state?` is written to the
uncommitted_changes field of the generated releases/ or dev_releases/
manifest file.  We don't care about this value either, but there's no
point submitting an upstream PR as they're rewriting bosh in Go.

The second time, it always calls