HCF-1212 Don't overwrite entrypoints in the SDL
We want to keep the default run.sh entrypoint from fissile, because that uses dumb-init correctly. Changing the entrypoint causes run.sh to stop working correctly.
This means that the custom env vars that we were loading as part of the entrypoint are now movin to scripts that get sourced as part of the normal startup of a fissile generated role.