HCF-1234 metron uses DNS instead of etcd to find doppler servers
It expects to have a separate round-robin DNS name for all doppler
servers in each AZ.

We don't have this, so we hard-code `spec.az` to "doppler-int" and
use the service domain suffix "hcf.svc" as the doppler address.  This
has to be done globally because virtually all roles have a
metron_agent job.

  "Fighting off both the sickness and the fear, Frodo gripped Sam's
  hand. 'Up!' he said in a hoarse breath without voice. 'It all comes
  from here, the stench and the peril.... Quick!'"

  -- The Two Towers, LoTR Book 4, Ch 9, Shelob's Lair