HATS 014: Force start the app after `cf restage`
Sometimes `cf restage` can fail to tail the logs, and finish before the
app has finished starting.  If we force a `cf start` afterwards, it will
do the right thing (and return success if the app was already started).
This avoids an issue where the app fails to start and the router instead
responds with a 404 app not found.