HATS: Allow customizing docker registry for test app
This makes it possible to run HATS on an air-gapped cluster by providing
a custom docker registry with the appropriate images.  By default this
is not set (and the main docker.io registry is used).

This does not allow all of HATS to pass due to the use of packagist.org
by composer when pushing PHP apps.