HCF-1234 Override zone for doppler and metron_agent jobs
The previous "fix" was overriding `spec.az` globally, which will conflict
with any future use of using it to actually specify the correct availability
zone.

This commit only overrides for the doppler and metron_agent jobs.  This still
does not properly implement the upstream intend of keeping logs of unaffected
zones available in the event of a zone failure. But that will require changes
to HCP to implement.