Remove invalid handling of MAKEFLAGS
MAKEFLAGS strips the leading dash (unless it starts with an option
that doesn't have a single-character shortcut), so is not useful for
explicit calls to `make`.

The whole construct only works when MAKEFLAGS is empty. When you
invoke `make -C dir releases` then MAKEFLAGS will be set to just `w`.
You can always run `make -j4 releases` if you want to build releases
in parallel.