Vagrant: Rebuild box to be kubernetes
The converts the packer build to create a vagrant image that runs
kubernetes directly, with no sharing of the host filesystem.  It can be
targetted at http://192.168.88.88:8080/ and requires no authentication,
and has support for the DNS addon.  It should be enough to run a
kube-ified SCF.

Based on OpenSUSE Leap 42.2 because that's our thing now.