Drop use of xxd in cert generation
This is a more inefficient passphrase generation method, but it's not
slower in practice (/dev/urandom is pretty fast), and drops the
requirement of having xxd available (which comes with vim).

Note that the LC_CTYPE change is necessary for macOS/BSD compatibility
(they assume actual text and don't accept random bytes as input).