make/kube: don't clobber FISSILE_OUTPUT_DIR, don't override it globally
Randomly deleting the global output directory just leads to having a
really bad time.  Always add the build target to it instead.

Also, this makes sure we do the hacky job grep thing on the directory
that actually contains our newly-created output, instead of assuming
that it's ${PWD}/${BUILD_TARGET}.