`unset CDPATH` in various scripts.
In some situations the envrionment variable CDPATH is set, causing
`cd` to print the name of the directory it changes into. This
unexpected output breaks constructions of the form
```
cd ... && pwd
```

Adding an early `unset CDPATH` prevents that now.

Note: While replacement of the above construct with `readlink -f ...`
was discussed it was decided against because of portability issues
with OSX/Darwin. While `readlink` itself is available on that
platform, `readlink -f` is not.

As part of the work two pre-existing uses of `readlink -f` were
removed as well, with a suitable rewrite of the command.