Packer: support building Jenkins slaves on Triton
Triton uses a different base image, so we need to add the vagrant user
and install helm separately. We also need Java for Jenkins.  Since it
requires a secondary disk for bulk storage, we need a bunch of systemd
units to line things up correctly to ensure we mount the hostpath stuff
on that disk.