Get rid of silly ${GIT_ROOT} in the Makefile
It is an eyesore, and seems to be cargo-culted anyways, as it can't possibly work:

    ${GIT_ROOT}/make/bosh-release src/diego-release
    ...
    ${GIT_ROOT}/bin/generate-dev-certs.sh cf bin/settings/certs.env

So only the script paths get the ${GIT_ROOT}/ prefix, but any filename arguments
don't. Which can only work if ${GIT_ROOT} is the current working directory, in which
case ${GIT_ROOT} becomes redundant (well, it could work if the scripts cd'ed into
the GIT_ROOT, but they don't).