8547RYHJ Add support for secure registries
make/registry will create a registry with basic auth and TLS support
if REGISTRY_USERNAME (and REGISTRY_PASSWORD) are set. It uses the
internalCA and the router_ssl cert (which includes the *.cf-dev.io
wildcard).

make/secure-registries will enforce secure registry checking and
installs the internalCA system wide. docker needs to be restarted both
to read the updated certs, and to process the modified DOCKER_OPTS.