dHhs91yG Bump to cf-release 278
All releases have been bumped to the version recommended in the 278 release
notes, except for cflinuxfs2-release, which has been bumped to the latest
version available right now.

The tcp_emitter job of the routing release no longer exists; it has been
merged with the regular route_emitter job of the diego-release. This job is
now co-located on the diego_cell so that local_mode can be enabled.

The tcp_emitter will hang indefinitely (like the tcp_router) if UAA isn't
fully working by the time it starts, so the new wait-for-uaa job provides
a pre-start script to implement this.