Don't run the secret-generator in a pre-install hook
Because of a Kubernetes issue of not being able to run a watch on a
Role, we can't run the secret-generator job as a pre-install hook. The
secret-generator Role wouldn't be available.

Instead, the secret-generator is run as a regular job, and all of the
other pods that depend on the Secret stay in a ContainerCreating state.