Make the cc-clock role wait until the api is ready
This prevents the cc-clock role from creating the schema_migrations
table when it tries to check for the existence of this table. Under
heavy load, it can be deadlocked trying to access this table, and fail
to find it. Afterwards, it will try to create the table, and could
deadlock the migration happening in the api role.

https://github.com/jeremyevans/sequel/blob/7995642c7fccdf1d8a8516cdf3d51db2859bb27e/lib/sequel/extensions/migration.rb#L768