Add a patch to remove timestamp check for rsyslog
In nats or doppler, only limited logs will be recorded to /var/log/messages,
then this file is no longer updated. If /var/log/messages is outdated for
more than 65 mins, there will be an alert named "timestamp failed".
So add a patch for these component to remove the check.