Move avahi-aliases job to router role
It needs to listen on NATS and cannot connect to the NATS_SERVICE_HOST
address because hairpin is no longer supported and NATS doesn't seem
to bind to the localhost interface.

The avahi-aliases job has also moved from the scf-helper-release to
scf-release.

Finally golang1.8 is no longer used in the scf-release, so switch to
golang1.10 instead.