autoscaler: Always use client_credentials
In the context of SCF (/other fissilized CF deployment), it is possible
for the administrator to change the administrator password and then do
something that requires the autoscaler to be restarted.  In that case
the password embedded in the helm release will no longer be correct,
which will cause the autoscaler to cease working.

Instead, use client credentials for the autoscaler (which would be
generated anyway).  This ensures that we always have an auth method that
will work even when the secrets are rotated.

This is independent of the autoscaler being used outside of SCF (where
it might still be a good idea to use username / password authentication)