Make sure file-server URL remains diego-access
This is the service name used by previous releases, and any running app will
have this URL cached in the Diego DesiredLRP, making it impossible for the
apps to remain running until they are explicitly restarted or restaged via the
cloud controller.

This fissile update also changes the stateful set selector to using the
skiff-role-name label, which is the selector autogenerated by previous
releases, that didn't define a selector themselves. This too is required
to have successful updates because changes to the pod spec of a stateful
set are not allowed during updates.

[trello#JmwDHOUp]