Prepare the ground for - New parameter for coming AZ code to pick up. - Service account to hook node reading permissions used by diego-cell into.
While the new service account is identical to `default` at the moment
having it separate from that allows us to attach the node reading permissions
later without giving the same to everybody else as well.

Note that fissile currently does not support the creation of cluster
roles and their bindings, as we would need to provide permissions for
nodes, a cluster-level resource. The user documentation will explain
how adding the permissions is done when it is needed.