autoscaler: move jobs to separate roles
On some kube systems hairpin connections (going to the same container
via the service) is not supported; as the operator connects to the
scheduler and the scalingengine this can be problematic.  Move it to a
new role that has nothing else to work around the issue.

Move all jobs to separate roles to avoid this issue.