Remove configure-HA-hosts.sh from roles that don't need it
The script sets KUBE_NATS_CLUSTER_IPS. KUBE_MYSQL_CLUSTER_IPS, and
KUBE_CONSUL_CLUSTER_IPS, but only the first one is still being used.

The validator was also checking for KUBERNETES_CLUSTER_DOMAIN, but
that one isn't being set by this script, it is set by run.sh in the
fissile stemcell.

The only remaining properties that require the script are:

  diego.route_emitter.nats.machines: ((KUBE_NATS_CLUSTER_IPS))((#KUBE_SIZING_NATS_COUNT))((/KUBE_SIZING_NATS_COUNT))
  nats.machines: ((KUBE_NATS_CLUSTER_IPS))((#KUBE_SIZING_NATS_COUNT))((/KUBE_SIZING_NATS_COUNT))

In addition to the obvious suspects, this is being used by all
roles that use the route_registrar.