Settings: use kube-internal DNS
The minibroker gives hosts of the form <blah>.cluster.local; that cannot
be resolved by world-visible DNS servers (because it's only a thing
within the kube cluster).  Ensure that we use the kube-provided DNS
server instead.

This of course depends on the kubernetes cluster setup, and should be
overridden by administrators for their specific cluster.