make: build *.tgz as well as *.zip for bundle archives
For the Jenkinsfile, we have to repeat the condition as Groovy gives
scoping errors (com.cloudbees.groovy.cps.impl.CpsCallableInvocation)
unless we scope the variables in an if block this way.  Also, the Groovy
docs appear to be incorrect, trying to match regular expressions in a
switch statement doesn't appear to work.