[#wqcKOPJC] cf bump. Bumped the releases to the set for cf-deployment 3.6.0
New releases, and ignored in this first step are
* bosh-dns-aliases-release	0.0.3
* bpm-release			0.12.2
* log-cache-release		1.4.3

Still ignored releases are
* cf-networking-release		2.14.0
* consul-release			195
* silk-release			2.14.0

The new garden-runc job
* netplugin-server

and the new loggregator jobs
* reverse_log_proxy_gateway_cf_auth_proxy
* reverse_log_proxy_gateway

are ignored as well.

A few property renames can be ignored, as we are using the renamed properties at all.
Most properties look to have sensible defaults.

Four new diego properties we may have to switch off, i.e. override their default of true
* bbs:set_kernel_parameters
* file_server:set_kernel_parameters
* locket:set_kernel_parameters
* rep:set_kernel_parameters

The relevant description is

	Enable tuning /proc/sys kernel parameters.
	NOTE: set this property to 'false' when
	deploying to BOSH-Lite or other containerized
	BOSH clouds.

While we are not bosh in any way, we are containerized.