[#4ceOnKG6] Make nfs-broker properly optional, i.e. not registered when off.
Move broker registration from `configure-scf` job to
new `register-nfs-persi`, as a post-start script. Run
the new job as colocated container of the broker to
separate the job and broker readiness. Running the new
job in the same container as the broker will generate
a dependency cycle preventing the pod from becoming ready

(   post-start waits on the endpoint to accept traffic before completing
 -> endpoint waits on pod to become ready before accepting traffic
 -> pod waits on post-start to complete before becoming ready
)