[#Cg5kMcEp] Convert the cc-clock's wait-for-api functionality from a patch to a pre-start script.
Note, the original code was modified here to select the port to talk
to based on the chosen protocol. We do not fiddle with the port
definitions and keep to the defaults. We further went for faster
checks, every 3 seconds, with an overall timeout of 1 hour (Vs. no
timeout in the original code).

Context and background:

Our switch to BPM since CAP 1.3.1 bypasses the old-style control
script modified by the patch. This disabled the bug fix the waiter
gives us (A database migration race condition between API and CC-CLOCK
under high load, dead-locking the cluster during deployment).

Further, having the waiter being a part of the control script was also
able to cause monit to incorrectly believe that the pod was healthy
while it was actually still waiting for API before starting the main
clock job. This missing job in turn allowed CC and Diego state to
desync, leading to leaked LRPs from re-pushing and scaling CF apps.