[#dOiHIcaP] Replaced the runtime installation of the SUSE userspace FUSE support reaching out to unknown machines, installing an unknown version with a combination of build- and runtime installation.
A new bosh package `suse-fuse` collects at build time the RPM archive
for the SUSE userspace FUSE support.

A new bosh job `install-suse-fuse` uses a pre-start script to install
the collected RPM archive, using options which prevent zypper from
reaching out to the network.

The name of the new job ensures that its pre-start script is run
before the pre-start script of the `mapfs` job which completes the
fuse setup.

Our patch for the `mapfs` pre-start script is modified to insert a
skip over the installation of the pre-packaged .deb archives when fuse
support is already installed, instead of inserting the old full
installation SUSE's fuse support via zypper and network.