[#dOiHIcaP] Dropped the runtime installation of the SUSE userspace `fuse` support. It was reaching out to unknown machines, installing an unknown version of `fuse` with a combination of build- and runtime installation. This also prevents air-gapped deployment.
The `fuse` support is now expected to be installed in the stemcell. IOW to be suitable for use by SCF a stemcell must have the `fuse` package installed.

Our patch for the `mapfs` pre-start script is modified to insert a skip over the installation of the pre-packaged .deb archives when `fuse`
support is already installed, instead of inserting the old full installation of SUSE's fuse support via zypper and network.