acceptance-tests-brain: syslog: Wait for ContainerCreating
It looks like sometimes we end up trying to get logs from the syslog
container while it's still in ContainerCreating, somehow (even though
all conditions are ready); this probably happens because there is no
conditions yet.

Explicitly check the container phase to work around this.

Also make it more obivous (in the error message) how to set up
SCF_LOG_HOST to get this to work.