brains/minibroker-mongo: disable pod liveness probe
On slow clusters, we could take too long to deploy the mongo pod; if
kube happens to kill the pod at a bad time, we could end up with an
unrecoverable state.

The proper fix is to change the mongodb helm chart to use an init
container, instead of just running the initial db setup in a script.

This also enables extra logging, in case that helps.