acceptance-tests-brain: wait for ready statefulsets better
When looking to see if StatefulSets are ready, we should fall back in
the case where the readyReplicas count doesn't exist yet.  Also,
actually check the _desired_ replica count, not the current replica
count, against the ready replica count.

Additionally add a line on when the global timer expires, to make debugging
post-mortem logs easier.