Fixed database with BPM persistent disk
Since the BPM persistent disk is mounted on /var/vcap/store/mysql, the
original script that runs the database setup stopped running, as the
creation of this directory itself was done as part of the pre-start
setup. The patch introduced in this commit fixes that by switching the
logic to a file inside that folder.