New script to attempt to recover stuck lock
This adds a script that attempts to recover a stale lock for the
autoscaler-metrics database.  This can occur if the autoscaler-metrics
pod was terminated while it's holding the lock, as for some unfathomable
reason the library implements locking by inserting data into a table,
instead of just using actual (possibly row-level) locking.