recovery-autoscaler-metrics-liquidbase-stale-lock: wait
This changes the script to automatically wait until we identify a dead
pod.  It is likely that kubernetes will automatically restart the pod
(unless it's already in CrashLoopBackOff), which makes it difficult for
the user to run the script at the exact time the error message is
available.  Instead, we will poll the logs in a loop until it shows up.