acceptance-tests-brain: replace SQL client packages
It's hard to install the postgres client binaries via zypper for
postgres (because the SLE stemcell doesn't have a good repository
setup); replace our use of those client binaries with a golang binary as
drivers are available for both MySQL and Postgres in pure-go versions.
This means we do not need to acquire separate drivers.

Also adding go1.12 as that makes dependency management easier (via
go module support).