Jenkinsfile: upload tested bundles separately
This makes us upload the archive bundles to a separate place before
running the tests, and then upload it again once the tests pass.  This
lets the QA machinery avoid testing builds that did not pass the tests
(but still allows us to examine the generated charts).

Fixes jsc#CAP-631