acceptance-tests-brain: cf-usb: document overrides
This puts more documentation on what methods on the base class should be
overridden by derived classes, in case we want to add even more tests in
the future.  Also moves some of the methods that only depend on the
sidecar type to the base class (keyed by the class name).

Addresses:
https://github.com/SUSE/scf/pull/2657#discussion_r321402654
https://github.com/SUSE/scf/pull/2657#discussion_r321403090
https://github.com/SUSE/scf/pull/2657#discussion_r321404236