Added support for external database server for CCDB &c.
This adds the option to use an external database server for the
CF-internal databases (CCDB, CF-USB, etc.), and UAA. This lets
administrators handle their own database backups if necessary.

A new job (found in the scf-helper-release) is used to initialize the
databases, using the same configuration that the mysql-release uses in
mariadb_ctrl.  The user must supply the external database host, port,
and credentials.

Currently this only supports MySQL/MariaDB databases.

Since the expression to calculate the correct internal database host,
port, etc.  is a bit complex, YAML anchors are used to avoid repeating
ourselves.

A feature guard is added to the internal database (proxy) roles,
enabling the user to manually disable them when using an external
database.

Note: The various routing roles (tcp-router, router, routing-api) used
the internal mysql service for the DNS health checks. This has been
changed to use the nats servicve instead, given that the mysql pods
can now be disabled by a feature flag (see previous paragraph).