Support external database server for CCDB &c.
This adds the option to use an external database server for the
CF-internal databases (CCDB, CF-USB, etc., and UAA when it is embedded
in the SCF deployment).  This lets administrators handle their own
database backups if necessary.

A new job is used to initialize the databases, using the same
configuration that the mysql-release uses in mariadb_ctrl.  The user
must supply the external database host, port, and credentials.

Currently this only supports MySQL/MariaDB databases.