Use kube csr feature for opi-registry in eirini
The old mechanism of using a docker cert copying daemon set does not
work with non-docker based container runtimes. It also has problems
when the host filesystem is not writable by the daemon set.

The new mechanism uses the kube certificate service to sign a cert for
the registry. This still requires the kube CA to be installed such that
the container runtimes on each worker node will trust it.

This commit uses a nodePort service on 127.0.0.1 for the registry:

* Can't use registry.$DOMAIN because it uses the wildcard cert from
  the gorouter. Also we don't want to expose the registry publicly.

* Can't use bits-bits-service.$NS.svc.cluster.local because the name
  doesn't resolve outside the cluster (container runtime cannot pull
  from there).

* Can't use localhost because it doesn't work (for unknown reasons).

All cluster-roles are moved out of the rolemanifest into chart-parts/
so that only the ones used in a specific configuration are defined
and bound to the service accounts.

[jsc#CAP-587]