CAP-587 Get rid of bits-service
We don't need it; we can just talk directly to the blobstore.

For Eirini we need the bits registry, but even Eirini's recipe
up- and downloaders will talk to blobstore if bit-service is not
enabled.

All this allows us to use just 127.0.0.1 as the common name for
the bits ssl cert. This is important on EKS because the kube cert
signer does not support subject alternate names, so the cert can
only have a single name, and it MUST be the registry address for
it to be trusted by the container runtime.

[jsc#CAP-587]