Enable BPM for bits-service
On CaaSP4 we often see the bits-0 pod hanging because monit does
not properly detect the running bitsgo process. Hopefully changing
to bpm will avoid this issue. Even if it doesn't, it is still a
useful change because now the bits-service will run under the vcap
user and not as root.

[jsc#CAP-748]