Changes to RC pipeline (#2720)
* Jenkinsfile: Always build RC builds

Turns out that even when we manually trigger builds, Jenkins tend to
pick up changes and then refuse to build (because more than the README
had changed).  It's much easier to disable polling on the RC branch in
Jenkins, and just always do manual triggers instead.

* Jenkinsfile: upload tested bundles separately

This makes us upload the archive bundles to a separate place before
running the tests, and then upload it again once the tests pass.  This
lets the QA machinery avoid testing builds that did not pass the tests
(but still allows us to examine the generated charts).

Fixes jsc#CAP-631

* make/include/versioning: stop dropping -rc1 suffix

We actually want that so the we can publish RC-labelled charts to the
helm repository.

* Jenkinsfile: fix calls to distSubDir

We were not passing in the new required argument in some call sites
(because they were never run during testing).