Use separate service account for configgin
Configgin needs to read/write secrets, which is not a capability
all pods should have. The configgin service account is passed
automatically (via fissile) to each container via the
CONFIGGIN_SA_TOKEN environment variable (from a secret maintained
by configgin-helper).

configgin-helper imports the MONIT_PASSWORD from the generated secrets,
so will not start until the secret-generator is done, so secret-generator
has to be bound to the configgin role too to avoid a deadlock.

Active/passive roles now get their own service account because they need
to modify annotations on pods and services (switchboard-leader).

garden-runc also has a separate service account because it needs to fetch
its own pod spec to determine which node it is running on to determine the
availability zone.

This commit also removes the secrets role from the eirini account; it is
unknown why that was needed in the past.

[jsc#CAP-828]