Vagrant: add hack to use obs://Kernel:stable
We are seeing issues with btrfs when running somewhat stressful
workloads (CATS, pushing apps in a loop); this appears to be resolved
via a newer kernel.  While the _correct_ way to do this is to rebuild
the vagrant image via packer, this is not feasible at this point in time
as we are having issues getting the Kubernetes images (since CaaSP 4 has
moved to using kubeadm and friends, and the old RPM images are no longer
available).

This is a temporary hack to update the kernel used when vagrant
provisions the VM instead; the more proper fix has been deferred to
jsc#CAP-1156.