scf-release: update golang to 1.13.5
Remove go1.10 and go1.12.9, and use (prebuilt) go1.13.5 instead.  This
ensures that we are not affected by CVE-2019-17596 (though we shouldn't
anyway, as all of the changes are to test code).

Note that this breaks the build due to old code (specifically,
github.com/docker/distribution#2697); this will be fixed in a follow-up
commit before this merges.