CAP-1125 Allow empty UAA_CA_CERT to use system cert store
Fixes #3187

One functional change: now unconditionally requires TLS for the connection
between credhub and the database. Commit d1059152 claims this is not supported
by our config, but it seems to work just fine with the credhub brain test.
This change is required to make credhub work with UAA_CA_CERT being empty and
the embedded UAA using the internal cert.

[jsc#CAP-1144]