acceptance-tests-brain: 010: show uploader logs on error
If the insecure-registry test fails, it would be nice to also display
logs from the uploader app, so that if _that_ has any clues on why
things failed we can get it out of the test logs.

This has no effect if the test passes (so they are kept concise), even
in verbose mode.